drm: avoid switching to text console if there is no panic timeout
Add a check for panic_timeout in the drm_fb_helper_panic() notifier: if
we're going to reboot immediately, the user will not be able to see the
messages anyway, and messing with the video mode may display artifacts,
and certainly get into several layers of complexity (including mutexes and
memory allocations) which we shall be much safer to avoid.

Note: This patch was recently reverted by the drm backport.

Sent upstream: https://lkml.org/lkml/2011/10/17/490

BUG=chromium-os:21607
TEST=platfrom_KernelErrorPaths

Change-Id: I3092f05feed5169fd05b2e206c59a2220b3c72da
Signed-off-by: Hugh Dickins <hughd@google.com>
[ Edited commit message and modified to short-circuit panic_timeout < 0
  instead of testing panic_timeout >= 0.  -Mandeep ]
Signed-off-by: Mandeep Singh Baines <msb@chromium.org>
Reviewed-on: http://gerrit.chromium.org/gerrit/10210
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>