ARM: tegra: timer: Fix us timer time after resume from LP1
In LP1 suspend, or during a failed LP0 suspend, the core power is
not turned off and the timer register is not reset.  Saving the
timer value during suspend and adding it to the offset after resume
will cause the timer value to double for each suspend cycle.

To fix, when resuming subtract the current value of the timer
register from the offset.  Also, use the rtc registers to add
the time that passed during suspend.

Change-Id: I9f7ee2089b98cf66af45271f8478fbd9eb2ee250
Signed-off-by: Colin Cross <ccross@android.com>