CHROMIUM: BACKPORT: pstore: new filesystem interface to platform persistent storage
Some platforms have a small amount of non-volatile storage that
can be used to store information useful to diagnose the cause of
a system crash.  This is the generic part of a file system interface
that presents information from the crash as a series of files in
/dev/pstore.  Once the information has been seen, the underlying
storage is freed by deleting the files.

BUG=chromium-os:19884
TEST=Build (disabled in config, no functional change)

Signed-off-by: Tony Luck <tony.luck@intel.com>
(cherry picked from commit ca01d6dd2d7a2652000307520777538740efc286)

Change-Id: I5e450af1c2aa78f2a2a43f9c5d9453640d2452d6
Reviewed-on: http://gerrit.chromium.org/gerrit/7023
Reviewed-by: Mandeep Singh Baines <msb@chromium.org>
Tested-by: Elly Jones <ellyjones@chromium.org>