CHROMIUM: arm: tegra: fix dependency on cache_l2x0
There was a dependency in suspend.c where CACHE_L2X0 was required to
be enabled. To remove this, the calls are wrapped with simple checks so
that the 2 calls to cache_l2x0 are only done if CACHE_L2X0 is defined
in the kernel config.

TEST=verified if CACHE_L2X0 is disabled, kernel builds, and that if
                 CACHE_L2X0 is enabled, the kernel uses the calls,
     verified kernel booted with new code and CACHE_L2X0 disabled.

BUG=chrome-os-partner:1946

Change-Id: Ibce232cdec0aa9c1e7d63fee0d2b54366318d859
Signed-off-by: Rhyland Klein <rklein@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6174006