drm/i915: Honour LVDS sync polarity from EDID
The i915 driver normally assumes the video bios has configured several
of the LVDS panel registers, and it just inherits the values. If the
vbios has not run, several of these will need to be setup. So we need to
check that the LVDS sync polarity is correctly configured per any
available modelines (e.g. EDID) and adjust if not, issuing a warning as
we do.

Signed-off-by: Mark Hayter <mdhayter@chromium.org>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
(cherry picked from commit aa9b500ddf1a6318e7cf8b1754696edddae86db9)
Change-Id: I8dc8daccb40c86e828fe41792d9b2971458a7a42
Reviewed-on: http://gerrit.chromium.org/gerrit/690
Reviewed-by: Olof Johansson <olofj@chromium.org>