CHROMIUM: i915: Backlight to PWM conversions round more accurately
Modified the formulas for backlight to PWM value and vice versa to round
to the nearest integer, rather than rounding down.  The old formula
caused problems with converting from backlight to PWM and then
converting back to backlight.  The result was not always the same as the
original value.

BUG=chromium-os:7615
TEST=Write to i915 backlight brightness and read actual_brightness.
Should match.

Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6264006

Reviewed-on: http://gerrit.chromium.org/gerrit/416
Reviewed-by: Benson Leung <bleung@chromium.org>
Tested-by: Olof Johansson <olofj@chromium.org>
Change-Id: I5d558ec835bb5db85be1f609ae72efdef9d21099
Reviewed-on: http://gerrit.chromium.org/gerrit/9977
Reviewed-by: Bryan Freed <bfreed@chromium.org>
Tested-by: Simon Que <sque@chromium.org>