BACKPORT: Bluetooth: ath3k: Avoid duplication of code
In commit 86e09287e4f8c81831b4d4118a48597565f0d21b, to reduce memory
usage, the functions of the ath3k module were rewritten to release the
firmware blob after it has been loaded (successfully or not).

The resuting code has some redundancy and the compiler can potentially
produce better code if we omit a function call that is unconditionally
executed in

,----
|     if (ath3k_load_firmware(udev, firmware)) {
|             release_firmware(firmware);
|             return -EIO;
|     }
|     release_firmware(firmware);
|
|     return 0;
| }
`----

It may also be argued that the rewritten code becomes easier to read,
and also to see the code coverage of the snippet in question.

BUG=chrome-os-partner:5499
TEST=hcitool scan

Signed-off-by: Rogério Brito <rbrito@ime.usp.br>
Cc: Alexander Holler <holler@ahsoftware.de>
Cc: "Gustavo F. Padovan" <padovan@profusion.mobi>
Cc: Miguel Ojeda <miguel.ojeda.sandonis@gmail.com>
Signed-off-by: Gustavo F. Padovan <padovan@profusion.mobi>
(cherry picked from commit 84f0e17f78471857104a20dfc57711409f68d7bf)

Change-Id: I92b240b809a3f1145587c806bf90ca9ad05e54fd
Signed-off-by: Jay Kim <yongjaek@google.com>
Reviewed-on: http://gerrit.chromium.org/gerrit/6677
Tested-by: Jay Kim <yongjaek@chromium.org>
Reviewed-by: Olof Johansson <olofj@chromium.org>