CHROMIUM: ASoC: tegra: Implement SPDIF CPU DAI
This is a minimal driver for the Tegra SPDIF controller.

In hardware, the SPDIF signal is always routed to any active HDMI display
controllers, and may also be routed to external pins on Tegra. This
minimal driver is targetted purely at HDMI usage, but may work for the
SPDIF pins too, given the correct pinmux settings.

BUG=
TEST=
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Olof Johansson <olof@lixom.net>

Review URL: http://codereview.chromium.org/6624031