CHROMIUM: arm: tegra: Fix backlight toggling on seaboard/fb0
Seaboard has both an internal panel (on fb0) and external HDMI (on
fb1).  One of the devices might be blanked while another is unblanked.
In order for the backlight blanking logic to know whether to turn off
the backlight for a particular framebuffer's blanking notification, it
needs to be able to check if a given framebuffer device corresponds to
the backlight.

This implement the check_fb pwm_backlight hook to make the internal
panel's backlight only toggle on fb0 blank changes.

BUG=chrome-os-partner:1812
TEST=HDMI on Seaboard, powering off HDMI does not also power off internal
backlight
Signed-off-by: Robert Morell <rmorell@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6624037