CHROMIUM: arm: tegra: seaboard: 1-kHz backlight PWM
Seaboard/Kaen was previously specified to 200 Hz PWM and driven by the
32 kHz clock.  However, the 32 kHz clock can only drive up to 128 Hz,
due to the 256 PWM duty cycle levels.

This patch selects a 12-MHz clock source instead, which can drive a
higher frequency.  It also sets the PWM frequency to 1 kHz so that it is
more within spec (200 Hz is at the low end of the frequency spec range).

BUG=chrome-os-partner:2117,chrome-os-partner:1957
TEST=Run on a tegra2 system, check backlight

Change-Id: I810178eea392ad6a4ccd44365f3bd1385e6d3ef8
Signed-off-by: Simon Que <sque@chromium.org>

Review URL: http://codereview.chromium.org/6777023