CHROMIUM: tegra: remove the clock sleepable WARNING
When boot the system, it will show WARNING:
can't make child clock apbdma of pclk sleepable if it's parent could change
The dvfs is trying to init clock sclk, it trickles down to apbdma.
Part of that init is to set cansleep flag. When setting a clock as cansleep,
in which case the entire list of clocks is traversed to set the children as
cansleep as well. in here, the clk_set_cansleep can only be called on clocks
that are the sole parent of all of their child clocks, meaning the child clock
can not be reparented onto a different, possibly non-sleeping, clock.
The apbdma has the set_parent callback, so the driver warns that this case is
not proper, but allows it anyways after.
But the apbdma is only has one parent, the flags for this clock is 0, not MUX,
it can not be reparented.
So we can remove this WARNING for !(flags & MUX).

BUG=none
TEST=test on seaboard, the WARNIN is gone.

Change-Id: Ica24b090653cba1886d60b1b266c9239d405630a

R=olofj@chromium.org,dahuang@nvidia.com
Signed-off-by: Wei Ni <wni@nvidia.com>
Signed-off-by: Olof Johansson <olof@lixom.net>

Review URL: http://codereview.chromium.org/6760022