CHROMIUM: ARM: Tegra: Harmony: Fix conflicting GPIO numbering
Currently, both the WM8903 and TPS6586x chips attempt to register with
gpiolib using the same GPIO numbers. This causes the gpiolib registration
for the second driver to fail.

To solve this, add a define to board-harmony.h for the TPS6586x, and make
board-harmony-power.c use this define, instead of directly referencing
TEGRA_NR_GPIOS.

Signed-off-by: Stephen Warren <swarren@nvidia.com>
Change-Id: I4db62583d818ccba3581a2f324525a564394b9ba
Reviewed-on: http://gerrit.chromium.org/gerrit/1212
Reviewed-by: Olof Johansson <olofj@chromium.org>