CHROMIUM: ARM: Tegra: Seaboard: Re-order sdhci device registration
The new order matches the order in kernel branch chromeos-2.6.37.

This is important because:

* U-Boot statically assigns MMC device IDs based on controller ID.
* U-Boot assumes that kernel MMC device ID numbering matches U-Boot numbering.
* U-Boot provides a kernel cmdline option e.g. root=/dev/mmcblk0p3 based on
  that numbering.
* The kernel dynamically assigns MMC device IDs based on enumeration order of
  the memory behind the host controller, rather than statically based on host
  controller ID like U-Boot.
* By registering the SDHCI controller for the built-in eMMC first, the
  enumeration of the built-in eMMC completes first (at least in practice), and
  hence eMMC gets assigned ID 0 just like U-Boot. If the SD slot is filled, it
  then gets assigned ID 1 just like U-Boot.
* If the MMC IDs mismatch, and the system boots from SD card not eMMC, the
  kernel will access the eMMC instead of SD card when attempting to mount
  /dev/mmcblk1p3 as the root fs. If eMMC is not partitioned/formatted, the
  kernel will panic since the root fs can't be mounted. If eMMC is partitioned
  and formatted, the kernel will mount an unexpected filesystem as the root fs.

Note that I'm not sure this is a 100% guaranteed solution to the problem; it
may just make the timing work out 99.9% of the time. Specifically, this change
makes eMMC controller registration happen first. I'm not 100% sure if this
guarantees that enumeration of the memory behind that controller is initiated
before other SDHCI controller registrations are executed. Equally, I doubt
this guarantees by design that enumeration of the eMMC will /complete/ first,
which is what actually affects ID assignment, although this does all appear to
be true in practice.

BUG=
TEST=Seaboard clamshell: Boot from SD card with eMMC blanked.
Signed-off-by: Stephen Warren <swarren@nvidia.com>
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6800019