CHROMIUM: ARM: tegra: kaen: add emc tables
This adds the EMC tables for kaen, selected by the ram id strappings on
the chip. We read these strappings and select the appropriate table.

It also adds some assertions to make sure that the functions aren't
called on the wrong machine since it risks destabilizing it.

Change-Id: I9fa44ad97e0fe86c87d96dd3627f5c49ecd47dd2
Signed-off-by: Danny Huang <dahuang@nvidia.com>
[olof: Changed to table based, minor tweaks]
Signed-off-by: Olof Johansson <olof@lixom.net>

BUG=chromium-os-partner:3253
TEST=Nvidia's dvfsstress testcase (yet to be delivered to us)

Review URL: http://codereview.chromium.org/6873117