CHROMIUM: arm: tegra: seaboard: Enable HDMI +5V for DDC
The HDMI specification says that +5V must be enabled when using DDC or
TMDS signals.  This change fills in the hook for Seaboard to set the
correct GPIO to provide this power during DDC transactions.

Additionally, initialize the +5V gpio to disabled, to save power until
it's needed.

BUG=chromium-os:1812
TEST=Test HDMI hotplug with Seaboard and Kaen, both before and after enabling HDMI with X11.
Signed-off-by: Robert Morell <rmorell@nvidia.com>
[olofj: 2.6.38 fixups]
Signed-off-by: Olof Johansson <olofj@chromium.org>

Review URL: http://codereview.chromium.org/6883112

Change-Id: I355ff5205ca453cc1550b7f73e01d14d56ac24a8