CHROMIUM: Tegra: temporary hack for setting suspend-to-RAM to LP1
On T25 based systems with BCTs that enable memory and core scaling,
resuming from suspend-to-RAM state results in a system reboot.  Setting
the suspend-to-RAM state to LP1 instead of LP0 allows the system to
resume successfully.

In order to allow testing to proceed, setting the suspend-to-RAM state
to LP1 as a temporary hack while we debug the real root cause.

BUG=chrome-os-partner:3388
TEST="cat /sys/kernel/debug/suspend_mode" and make sure it reads lp1.
Then put the system into suspend-to-RAM state followed by a successful wake.

Change-Id: I59bbde3e258903e98583b40e684c218f60bce1b3
Signed-off-by: Sameer Nanda <snanda@chromium.org>

Review URL: http://codereview.chromium.org/6889011